+++ src/test/java/step3/domain/RandomRacingRuleTest.java
@@ -2,6 +2,7 @@ package step3.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.service.RandomRacingRule;
 
 import static org.assertj.core.api.Assertions.assertThat;
